+++ wled00/FX_fcn.cpp
@@ -106,20 +106,25 @@ void WS2812FX::setPixelColor(uint16_t n, uint32_t c) {
   setPixelColor(n, r, g, b, w);
 }
 
+#define REV(i) (_length - 1 - (i))
+
+//used to map from segment index to physical pixel, taking into account grouping, offsets, reverse and mirroring
 uint16_t WS2812FX::realPixelIndex(uint16_t i) {
   int16_t iGroup = i * SEGMENT.groupLength();
 
   /* reverse just an individual segment */
   int16_t realIndex = iGroup;
-  if (IS_REVERSE)
-    if (IS_MIRROR)
-      realIndex = SEGMENT.length() / 2 - iGroup - 1;  //only need to index half the pixels
-    else
+  if (IS_REVERSE) {
+    if (IS_MIRROR) {
+      realIndex = (SEGMENT.length() -1) / 2 - iGroup;  //only need to index half the pixels
+    } else {
       realIndex = SEGMENT.length() - iGroup - 1;
+    }
+  }
 
   realIndex += SEGMENT.start;
   /* Reverse the whole string */
-  if (reverseMode) realIndex = _length - 1 - realIndex;
+  if (reverseMode) realIndex = REV(realIndex);
 
   return realIndex;
 }
@@ -176,18 +181,23 @@ void WS2812FX::setPixelColor(uint16_t i, byte r, byte g, byte b, byte w)
     for (uint16_t j = 0; j < SEGMENT.grouping; j++) {
       int16_t indexSet = realIndex + (reversed ? -j : j);
       int16_t indexSetRev = indexSet;
-      if (reverseMode) indexSetRev = _length - 1 - indexSet;
+      if (reverseMode) indexSetRev = REV(indexSet);
       #ifdef WLED_CUSTOM_LED_MAPPING
       if (indexSet < customMappingSize) indexSet = customMappingTable[indexSet];
       #endif
       if (indexSetRev >= SEGMENT.start && indexSetRev < SEGMENT.stop) {
         bus->SetPixelColor(indexSet + skip, col);
-        if (IS_MIRROR)  //set the corresponding mirrored pixel
-          bus->SetPixelColor(SEGMENT.stop - (indexSet + skip) + SEGMENT.start - 1, col);
+        if (IS_MIRROR) { //set the corresponding mirrored pixel
+          if (reverseMode) {
+            bus->SetPixelColor(REV(SEGMENT.start) - indexSet + skip + REV(SEGMENT.stop) + 1, col);
+          } else {
+            bus->SetPixelColor(SEGMENT.stop - indexSet + skip + SEGMENT.start - 1, col);
+          }
+        }
       }
     }
   } else { //live data, etc.
-    if (reverseMode) i = _length - 1 - i;
+    if (reverseMode) i = REV(i);
     #ifdef WLED_CUSTOM_LED_MAPPING
     if (i < customMappingSize) i = customMappingTable[i];
     #endif